Fix Decoder compression support
In Sereal V1 offsets are counted from the start of the document, but for
documents using the '2' encoding, it does not take into account the compressed
length (I'm not even sure protocol V1 supported encoding type '2', but
TestSet.pm produces it).

I'm not sure why the '-1' was there in the first place.