Add a generic type mapping interface for decoding
Basic usage allows customizing which Java class is used when decoding a Sereal
array/hash (defaults to ArrayList/HashMap), and there is also an hook to
customize object mapping.

This commit also removes ObjectType and the code to dynamically compile a
bean-like class for objects. It's not coming back. Ever.

A followup commit will add a diffferent way for mapping Sereal objects to Java
objects.