Fix issue #128, sparse arrays can cause segfault
Dumping a sparsely populated magic array in older perls can result
in dereferencing NULL, throwing a segv. For implementation reasons
this does not happen for non-magic perls, and apparently newer perls
stores something other than NULL for empty array slots.

This fixes the problem, but needs to be followed up with tests, etc.