Perl: 5.25 compatibility changes around op_sibling
In Perl 5.25, the OP structure's op_sibling member was renamed to allow
for the "parent op" tracking functionality. This means that any access
has to go through macros.

This workaround is a bit cludgy for pre-5.25 compatibility. This should
probably live in Devel::PPPort instead (but it's not clear how to do
this interface wise since old perls can't track parents at all).
I hope I didn't mess this up.