Perl/Decoder: check for integer overflow in SRL_RDR_ASSERT_SPACE()
ptrdiff_t is a signed type, but in a number of places we're passing
lengths which are unsigned.  If a huge varint is read, when converted to
a ptrdiff_t it will be negative.  The check asserts that 'len' is less
than the remaining space.  The remaining space is always >= 0, and if
len is negative then the test is always false and the so the check for
'enough remaining space' succeeds, and we proceed to read past the
buffer.