Improving the Python decoder for Sereal
Adding support for:

* Double (very trivial to add)
* Object (as a workaround, it returns {'class': 'classname', 'object', ...}
* RegExp (Perl regexes are different than Python's, but the code ignores this and just creates a regex anyway)

Also added support for version 4 of the protocol.

Also reading (and discarding) the optional header suffix.

Also removed a bunch of parentheses in `if` statements, as those are not needed.

Also reworked the code to allow anything to be tracked as a reference.
However, references to immutable objects are not allowed (they're
immutable after all), so such objects are just copied. Still, I believe
this is better than just a cryptic error message.

The tracking code is duplicated in too many places, it's probably a
better idea to split the _decode_bytes() function in two, so that we
don't have to repeat self._track_item all the time.