py: Fix auto decoding BINARY tags as UTF-8 (#214)
According the the Sereal spec BINARY and SHORT BINARY tags should be read
verbatim. In the first release of the Python decoder BINARY tags were
treated as STR_UTF8. This commit fixes this behaviour to comply with the
decoding spec. See the README details for changing the decoder behaviour.

WARNING: This is a breaking commit for Python 2 users as Python 2
will no longer be supported. See Python 2 deprecation for more details.