Moved bin_mode_classic from **kwargs into a proper named parameter
Follow-up of 49890e7fe415987c60588b3551f277113b76e869 and
df33d1c458d3baeb3c34ef319f30b940688a1964.

Python has named parameters, I don't think we need to pass an extra dict
with settings. Just pass it as a normal parameter. Also, just store that
as a plain attribute into this instance, which is simpler to write,
simpler to read, and simpler to understand.

In addition, this commit will make the Python interpreter throw an error
if an invalid parameter is passed, instead of silently ignoring it.