Decoder: build process - use diag to show regexp type not pragma message
gcc's pragma message is broken and shows far more than it should, which
is annoying. We used to use it to show the type of regexp representation
perl uses, but it makes the output very ugly and can fool people into
thinking it is broken, so instead we expose this as an XS routine
and use diag to show the value.