Encoder.xs: Fixups for op-stitching code
Also some whitespace tweaks to Decoder.xs so that the two are the same
for this piece of logic.

In 041e46b4f8f6948b568ccf31e1fca44d28cb4b13 I ported changed made
by Dave M to the Decoder into the Encoder. Unfortunately in the process
of making the ported patch backwards compatible with Perl's lacking
op_sibling_splice() I mistakenly put one statement on the wrong side
of a #if/#endif, which meant it did not work correctly (even though
it compiled fine.)

To add insult to injury we weren't testing the custom opcode
implementation properly, with this result this went unnoticed until
BBC (Blead Breaks CPAN) tests showed that MaxMind-DB-Writer-0.300003
was broken and also revealed that Data-Serializer-Sereal-1.05 was
also broken, both pointing at issues with Sereal::Encoder via
sereal_encoder_with_object which is implemented as a custom opcode.

This patch fixes all of this, and adds tests 190_customop.t to
both packages to ensure that this is tested properly moving forward.

Thanks to Dave M, Slaven Rezić, Tony Cook, Sergey Aleynikov for
playing a role in getting this identified and fixed.

This patch should mean that https://github.com/Perl/perl5/issues/17245
can be closed.