Go: proper support for json.Number values
When unmarshalling free-form json in Go, numbers become either float64 or json.Number. With the first option you lose precision for large integers, so it makes sense to always use the second.

But in current implementation of sereal/encode.go the json.Number values is serialized as strings. So, if you do a conversion like this:

json -> data -> sereal -> data -> json,

all numbers in your structure will become strings, which is not good.

Since encoding/json is a core module, and interoperability with it is a practical issue, it makes sense to support proper serialization of json.Number right here, which this commit does.