Hardening XS against ZipBomb for Snappy, zlib & zstd
Simply refuse to decompress if uncompressed data is bigger than `max_uncompressed_size`.
It works even if the attacker spoofs the uncompressed data size, because
Snappy, zlib & zstd are all robust enough to abort the decompression
after consuming the expected amount of data when there is still data to
decompress.