update miniz to 3.0.2
This adds support to the Makefile.PL's for the Encoder and Decoder
so that the env var USE_UNALIGNED can be used to control if miniz
uses unaligned stores and loads.

If USE_UNALIGNED is 1 or "yes" then we will build miniz with support
for unaligned loads and stores. It defaults to not doing so. The
CPAN mode build will prompt() the user if they want to do this if
the env var is not set.