update zstd to 1.5.2
This also includes some necessary and some nice-to-have improvements
to zstd/Makefle.PL. We now have support for building zstd with the
assembly code. I ended up hand rolling some of the Makefile logic to
avoid unnecessary rebuilds.

By default we will build with ZSTD assembly code for performance.
The NO_ASM or ZSTD_DISABLE_ASM env var can be set to a true value
to disable this, and the Makefile.PL should prompt() the user to
use it or not when neither env var is defined.

See https://github.com/facebook/zstd/blob/dev/lib/Makefile for the
original Makefile that ours is derived from.