go: add decoder option to treat REFP more similarly to REFN
Before this change:

    v1, v2 := 1, 1
    a1 := [&v1, &v2]
    a2 := [&v1, &v1]

    unmarshal(marshal(a1)) -> [1, 1]
    unmarshal(marshal(a2)) -> [1, &1]

While this preserves referential integrity for all values, it makes the decoded
data structure harder to use.

This commit adds an option to make both cases above round-trip to [1, 1]. This
makes the resulting structure easier to manipulate at the small price of
breaking referential integrity for scalar values.